perf(PageHandler): Remove the `VideoContainer` abstraction + Rewrite `Video` and `Channel` objects + Move files around and rename
Removing the `VideoContainer` abstraction because it was inefficient to fetch those containers and observe mutations on them. A lot of events that was flooding down the chain. It is replaced with dynamic `querySelectorAll` calls that produce a much smaller array to iterate through and grab only the new video containers we are interested in.